+++ src/components/app/MainMenuComponent.vue
@@ -1,29 +1,40 @@
 <template>
   <div class="menu">
     <v-navigation-drawer permanent :mini-variant="retracted">
-      <v-btn text @click="toggleRetracted">
-        <v-icon v-if="retracted">mdi-chevron-right</v-icon>
-        <v-icon v-else>mdi-chevron-left</v-icon>
-        <span v-if="!retracted">Hide menu</span>
-      </v-btn>
-      <v-list
-        nav
-      >
+      <v-tooltip right 
+        :disabled="!retracted">
+        <template v-slot:activator="{on}">
+          <v-btn text @click="toggleRetracted" v-on="on">
+            <v-icon v-if="retracted">mdi-chevron-right</v-icon>
+            <v-icon v-else>mdi-chevron-left</v-icon>
+            <span v-if="!retracted">Hide menu</span>
+          </v-btn>
+        </template>
+        Expand menu
+      </v-tooltip>
+      <v-list nav>
         <v-list-item-group color="primary">
-          <v-list-item
+          <v-tooltip right 
             v-for="menu in menus"
             :key="menu.title"
-            link
-            :to="menu.path"
-          >
-            <v-list-item-icon>
-              <v-icon>{{ menu.icon }}</v-icon>
-            </v-list-item-icon>
+            :disabled="!retracted">
+            <template v-slot:activator="{on}">
+              <v-list-item
+                v-on="on"
+                link
+                :to="menu.path"
+              >
+                <v-list-item-icon>
+                  <v-icon>{{ menu.icon }}</v-icon>
+                </v-list-item-icon>
 
-            <v-list-item-content>
-              <v-list-item-title>{{ menu.title }}</v-list-item-title>
-            </v-list-item-content>
-          </v-list-item>
+                <v-list-item-content>
+                  <v-list-item-title>{{ menu.title }}</v-list-item-title>
+                </v-list-item-content>
+              </v-list-item>
+            </template>
+            {{menu.title}}
+          </v-tooltip>
         </v-list-item-group>
       </v-list>
     </v-navigation-drawer>